dump data and fix some bug

+++ backend/database/migrations/2023_10_16_072936_create_orders_table.php
@@ -15,8 +15,8 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->integer('product_id');
-            $table->integer('user_id');
+            $table->unsignedBigInteger('product_id');
+            $table->unsignedBigInteger('user_id');
             $table->integer('quantity');
             $table->text('message');
             $table->timestamps();
